+++ pref/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.1.0
 
 - Add `reversed` attributes to boolean widgets
+- Improve dropdown layout
 
 ## 1.0.0